third_party/blink/web_tests/platform/mac/external/wpt/css/css-text/i18n/zh/css-text-line-break-zh-pr-strict-expected.txt

This is a testharness.js-based test.
[FAIL] 00B1  PLUS-MINUS SIGN may NOT appear at line start if zh and strict
  assert_approx_equals: expected 91.625 +/- 1 but got 61.625
[FAIL] 20AC  EURO SIGN may NOT appear at line start if zh and strict
  assert_approx_equals: expected 91.59375 +/- 1 but got 61.59375
[FAIL] 2116  NUMERO SIGN may NOT appear at line start if zh and strict
  assert_approx_equals: expected 99.75 +/- 1 but got 69.75
[FAIL] FE69  SMALL DOLLAR SIGN may NOT appear at line start if zh and strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] FF04  FULLWIDTH DOLLAR SIGN may NOT appear at line start if zh and strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] FFE1  FULLWIDTH POUND SIGN may NOT appear at line start if zh and strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] FFE5  FULLWIDTH YEN SIGN may NOT appear at line start if zh and strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] FFE6  FULLWIDTH WON SIGN may NOT appear at line start if zh and strict
  assert_approx_equals: expected 99 +/- 1 but got 69
Harness: the test ran to completion.
